Create README.md
The written code helps to find the historical CHF sales value (Ask) valid on a given day.
I used the API for this tool. To find a specific date, change the date "2014-12-30" to the date you are searching for.
The code finds the CHF selling rate applicable on a given day in Poland according to the Polish national bank.